Move auth stuff into separate module
